feat(activerecord): fix generator gaps A-D and materialize declares for 6 remaining models (RFC 0019) (#4222)
## Summary

Closes the 4 generator gaps blocking the final 6 skipped models, then
bakes their `declare` members, burning the RFC 0019 exclude list to
zero.

**Gap A** — Ruby `Module::Class` namespaced `className` strings emit
invalid TS identifiers. Fix: `buildNamespacedClassRegistry` scans all
model files for `static moduleName`/`static _demodulizedName` and builds
a lookup used by `buildExtendedAliases` to resolve
`"Namespaced::Client"` → `"Client"` (with `"Base"` fallback). Covers
`company.ts`, `company-in-module.ts`.

**Gap B** — `classify("iris")` → `"Iri"` (wrong singularization). Fix:
after `classify` produces a name absent from the model registry, try
`camelize` (no singularization); if that hits, add a `classified →
camelized` alias entry. Covers `eye.ts`.

**Gap C** — Subclasses that re-declare a hasMany/HABTM inherited from a
parent cause `TS2416` (`AssociationProxy<T>` is invariant). Fix:
`collectConflictingCollections` in `synthesize.ts` detects and
suppresses those overrides. A second variant: `dirty.test.ts` extends
`Pirate` and overrides `catchphrase` as a get/set accessor (`TS2611` —
TypeScript forbids overriding a `declare` property as an accessor). Fix:
add `@trails-typegen skip-columns: catchphrase` JSDoc to `Pirate` so the
schema column is not declared; the accessor in the test subclass then
works without `override`. Covers `cpk.ts`, `pirate.ts`.

**Gap D** — `composedOf` aggregation columns were declared as bare
scalars. Fix: `extractComposedOfColumns` scans AST for `composedOf`
mapping arrays and passes the result as `composedOfColumns` to the
synthesizer, which skips those columns. Covers `customer.ts` (all 6
columns mapped → file unchanged).

`user-with-invalid-relation` remains as documented permanent skip
(hand-written association that references a missing model).

## LOC note

Diff is ~565 LOC (after prettier: ~657), of which ~313 LOC are
machine-generated `declare` lines in model files. The infrastructure
changes are ~250 LOC. This exceeds the 500 LOC ceiling; the excess is
entirely machine-generated output with no logic to review.

## Test plan

- [x] `node scripts/typecheck.mjs` — passes with zero errors
- [x] Virtualizer unit tests (`virtualize.test.ts`,
`resolve-target.test.ts`) — 81/81 pass
- [x] `singular-reader-stale-target.test.ts` — 2/2 pass (required cast
of `anotherFirmId` after `client_of` gained a typed declare)
- [x] Materializer idempotent: re-running produces no changes on
already-materialized files

Closes-story: materialize-declares-rollout-final-six

+++ packages/activerecord/src/associations/singular-reader-stale-target.test.ts
@@ -47,7 +47,7 @@ describe("BelongsToAssociationsTest", () => {
     expect(await targetId(client.firm)).toBe(firstFirmId);
     expect(await targetId(client.firmWithCondition)).toBe(firstFirmId);
 
-    client.client_of = anotherFirmId;
+    client.client_of = anotherFirmId as bigint;
 
     expect(firmProxy.isStaleTarget()).toBe(true);
     expect(firmWithConditionProxy.isStaleTarget()).toBe(true);
@@ -66,7 +66,7 @@ describe("BelongsToAssociationsTest", () => {
     // Changing the owner FK makes the loaded target stale. Rails' guard
     // `(@stale_state && stale_target?) || find_target?` issues a fresh query
     // on the stale branch rather than returning the in-memory cache.
-    client.client_of = anotherFirmId;
+    client.client_of = anotherFirmId as bigint;
 
     expect(firmProxy.isStaleTarget()).toBe(true);
     expect(await targetId(firmProxy.loadTarget())).toBe(anotherFirmId);
